+++ src/main/java/com/example/solidconnection/application/controller/ApplicationController.java
@@ -6,7 +6,7 @@ import com.example.solidconnection.application.dto.ApplyRequest;
 import com.example.solidconnection.application.service.ApplicationQueryService;
 import com.example.solidconnection.application.service.ApplicationSubmissionService;
 import com.example.solidconnection.custom.resolver.AuthorizedUser;
-import com.example.solidconnection.custom.security.annotation.RequireAdminAccess;
+import com.example.solidconnection.security.annotation.RequireAdminAccess;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
