+++ src/main/java/com/example/solidconnection/custom/resolver/AuthorizedUserResolver.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.custom.resolver;
 
 import com.example.solidconnection.custom.exception.CustomException;
-import com.example.solidconnection.custom.security.userdetails.SiteUserDetails;
+import com.example.solidconnection.security.userdetails.SiteUserDetails;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
