--- src/main/java/com/example/solidconnection/custom/security/annotation/RequireAdminAccess.java
@@ -1,11 +1,0 @@
-package com.example.solidconnection.custom.security.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface RequireAdminAccess {
-}
