+++ src/main/java/com/example/solidconnection/security/annotation/RequireAdminAccess.java
@@ -1,0 +1,11 @@
+package com.example.solidconnection.security.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface RequireAdminAccess {
+}
