+++ src/main/java/com/example/solidconnection/security/filter/JwtAuthenticationFilter.java
@@ -1,0 +1,48 @@
+package com.example.solidconnection.security.filter;
+
+import com.example.solidconnection.security.authentication.JwtAuthentication;
+import com.example.solidconnection.security.authentication.SiteUserAuthentication;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+import static com.example.solidconnection.util.JwtUtils.parseTokenFromRequest;
+
+
+@Component
+@RequiredArgsConstructor
+public class JwtAuthenticationFilter extends OncePerRequestFilter {
+
+    private final AuthenticationManager authenticationManager;
+
+    @Override
+    public void doFilterInternal(@NonNull HttpServletRequest request,
+                                 @NonNull HttpServletResponse response,
+                                 @NonNull FilterChain filterChain) throws ServletException, IOException {
+        String token = parseTokenFromRequest(request);
+        if (token == null) {
+            filterChain.doFilter(request, response);
+            return;
+        }
+
+        JwtAuthentication authToken = createAuthentication(token);
+        Authentication auth = authenticationManager.authenticate(authToken);
+        SecurityContextHolder.getContext().setAuthentication(auth);
+
+        filterChain.doFilter(request, response);
+    }
+
+    private JwtAuthentication createAuthentication(String token) {
+        return new SiteUserAuthentication(token);
+    }
+}
